share notifiers (WIP)
migrate notifiers to new index from _moloch_shared user (WIP)
create new notifiers component for settings page
